refactor(BA-6919): relocate EntityType/ScopeType to common.data.entity
Move the open EntityType/ScopeType NewTypes and the EntityRef/ScopeRef
value objects from common.entity to common.data.entity, aligning them
with the sibling common.data.permission package. The module is a leaf
in common (depends only on common.identifier), so relocating within
common introduces no import cycle and no cross-package layering change.
Drop the package-level __init__ re-export; callers already import the
types module directly.
