+++ src/ai/backend/common/data/permission/types.py
@@ -81,7 +81,7 @@ class OperationType(enum.StrEnum):
 class EntityType(enum.StrEnum):
     """
     Deprecated for RBAC: use ``RBACElementType`` instead
-    or use ai.backend.common.entity.types.EntityType for non-RBAC-specific contexts.
+    or use ai.backend.common.data.entity.types.EntityType for non-RBAC-specific contexts.
     """
 
     # === RBAC scope/resource types (original 12) ===
