--- src/ai/backend/common/entity/__init__.py
@@ -1,9 +1,0 @@
-from .types import (
-    EntityType,
-    ScopeType,
-)
-
-__all__ = (
-    "EntityType",
-    "ScopeType",
-)
