+++ src/ai/backend/manager/actions/single_entity/base.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 
+from ai.backend.common.data.entity.types import EntityType
 from ai.backend.common.data.permission.types import Permission
-from ai.backend.common.entity.types import EntityType
 from ai.backend.common.identifier.entity import EntityID
