+++ src/ai/backend/manager/data/permission/virtual_scope.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
-from ai.backend.common.entity.types import EntityRef
+from ai.backend.common.data.entity.types import EntityRef
 from ai.backend.common.identifier.user import UserID
