+++ src/ai/backend/manager/models/virtual_scope/entity_membership.py
@@ -5,9 +5,9 @@ from datetime import datetime
 import sqlalchemy as sa
 from sqlalchemy.orm import Mapped, mapped_column
 
+from ai.backend.common.data.entity.types import EntityType
 from ai.backend.common.data.permission.types import Permission
 from ai.backend.common.data.permission.virtual_scope import EntityMembershipData
-from ai.backend.common.entity.types import EntityType
 from ai.backend.common.identifier.entity import EntityID
 from ai.backend.common.identifier.virtual_scope import VirtualScopeID
 from ai.backend.manager.models.base import (
