+++ src/ai/backend/manager/models/virtual_scope/scope_binding.py
@@ -5,9 +5,9 @@ from datetime import datetime
 import sqlalchemy as sa
 from sqlalchemy.orm import Mapped, mapped_column
 
+from ai.backend.common.data.entity.types import ScopeType
 from ai.backend.common.data.permission.types import Permission
 from ai.backend.common.data.permission.virtual_scope import ScopeBindingData
-from ai.backend.common.entity.types import ScopeType
 from ai.backend.common.identifier.scope import ScopeID
 from ai.backend.common.identifier.virtual_scope import VirtualScopeID
 from ai.backend.manager.models.base import (
